Implementing Event Dispatcher

+++ Source/Graphics/ARenderer.h
@@ -6,7 +6,7 @@ class ARenderer
 public:
     virtual void initialize() = 0;
     virtual bool ready() = 0;
-    virtual bool openWindow(const char* title, float width, float height) = 0;
+    virtual bool openWindow(const char* title, int width, int height) = 0;
     virtual bool destroyWindow() = 0;
     virtual void beginDraw() = 0;
     virtual void endDraw() = 0;
